Fixed fake value signatures in test decoder

+++ tests/test_hpack_decoder.c
@@ -27,10 +27,10 @@ FAKE_VALUE_FUNC(uint32_t, hpack_utils_encoded_integer_size, uint32_t, uint8_t);
 FAKE_VALUE_FUNC(int8_t, hpack_tables_static_find_name_and_value, uint8_t, char *, char *);
 FAKE_VALUE_FUNC(int8_t, hpack_tables_static_find_name, uint8_t, char *);
 FAKE_VALUE_FUNC(uint32_t, hpack_tables_get_table_length, uint32_t);
-FAKE_VALUE_FUNC(int, hpack_tables_init_dynamic_table, hpack_dynamic_table_t *, uint32_t, header_pair_t * );
-FAKE_VALUE_FUNC(int, hpack_tables_dynamic_table_add_entry, hpack_dynamic_table_t *, char *, char *);
-FAKE_VALUE_FUNC(int, hpack_tables_find_entry_name_and_value, hpack_dynamic_table_t *, uint32_t, char *, char *);
-FAKE_VALUE_FUNC(int, hpack_tables_find_entry_name, hpack_dynamic_table_t *, uint32_t, char *);
+FAKE_VALUE_FUNC(int8_t, hpack_tables_init_dynamic_table, hpack_dynamic_table_t *, uint32_t, header_pair_t * );
+FAKE_VALUE_FUNC(int8_t, hpack_tables_dynamic_table_add_entry, hpack_dynamic_table_t *, char *, char *);
+FAKE_VALUE_FUNC(int8_t, hpack_tables_find_entry_name_and_value, hpack_dynamic_table_t *, uint32_t, char *, char *);
+FAKE_VALUE_FUNC(int8_t, hpack_tables_find_entry_name, hpack_dynamic_table_t *, uint32_t, char *);
 
 /* List of fakes used by this unit tester */
 #define FFF_FAKES_LIST(FAKE)                     \
